- Moved the singleton class into its own header
